excel-backend/requirements.txt
fastapi==0.115.12
uvicorn==0.23.2
python-multipart==0.0.6
pandas==2.1.1
openpyxl==3.1.2
numpy==1.26.0
jinja2==3.1.2
motor==3.3.1
pymongo==4.6.0
dnspython==2.4.2
python-jose==3.3.0
passlib==1.7.4
pydantic>=1.7.4,<2.0.0